+++ R/S3-methods.R
@@ -23,5 +23,10 @@ print.lbl_df <- function(x, ..., n = NULL, width = NULL) {
 
 #' @export
 print.sjmisc.frq <- function(x, ...) {
+  # get variable label
+  lab <- attr(x$mydat, "label", exact = T)
+  # print label
+  if (!is.null(lab)) cat(sprintf("# %s\n\n", lab))
+  # print frq-table
   print.data.frame(x$mydat, ..., row.names = FALSE)
 }